+++ src/Task/BuiltIn/Composer/InstallTask.php
@@ -11,14 +11,13 @@
 namespace Mage\Task\BuiltIn\Composer;
 
 use Symfony\Component\Process\Process;
-use Mage\Task\AbstractTask;
 
 /**
  * Composer Task - Install Vendors
  *
  * @author Andrés Montañez <andresmontanez@gmail.com>
  */
-class InstallTask extends AbstractTask
+class InstallTask extends AbstractComposerTask
 {
     public function getName()
     {
@@ -36,19 +35,13 @@ class InstallTask extends AbstractTask
         $cmd = sprintf('%s install %s', $options['path'], $options['flags']);
 
         /** @var Process $process */
-        $process = $this->runtime->runCommand(trim($cmd));
+        $process = $this->runtime->runCommand(trim($cmd), $options['timeout']);
 
         return $process->isSuccessful();
     }
 
-    protected function getOptions()
+    protected function getComposerOptions()
     {
-        $options = array_merge(
-            ['path' => 'composer', 'flags' => '--optimize-autoloader'],
-            $this->runtime->getMergedOption('composer'),
-            $this->options
-        );
-
-        return $options;
+        return ['flags' => '--optimize-autoloader', 'timeout' => 120];
     }
 }