+++ src/Task/BuiltIn/Deploy/RsyncTask.php
@@ -45,7 +45,8 @@ class RsyncTask extends AbstractTask
         }
 
         $excludes = $this->getExcludes();
-        $cmdRsync = sprintf('rsync -e "ssh -p %d %s" %s %s ./ %s@%s:%s', $sshConfig['port'], $sshConfig['flags'], $flags, $excludes, $user, $host, $targetDir);
+        $from = $this->runtime->getEnvOption('from', './');
+        $cmdRsync = sprintf('rsync -e "ssh -p %d %s" %s %s %s %s@%s:%s', $sshConfig['port'], $sshConfig['flags'], $flags, $excludes, $from, $user, $host, $targetDir);
 
         /** @var Process $process */
         $process = $this->runtime->runLocalCommand($cmdRsync, 600);
@@ -54,7 +55,7 @@ class RsyncTask extends AbstractTask
 
     protected function getExcludes()
     {
-        $excludes = $this->runtime->getEnvOption('exclude', []);
+        $excludes = $this->runtime->getMergedOption('exclude', []);
         $excludes = array_merge(['.git'], array_filter($excludes));
 
         foreach ($excludes as &$exclude) {