+++ src/Task/BuiltIn/Deploy/Tar/PrepareTask.php
@@ -42,7 +42,8 @@ class PrepareTask extends AbstractTask
 
         $excludes = $this->getExcludes();
         $flags = $this->runtime->getEnvOption('tar_create', 'cfzp');
-        $cmdTar = sprintf('tar %s %s %s ./', $flags, $tarLocal, $excludes);
+        $from = $this->runtime->getEnvOption('from', './');
+        $cmdTar = sprintf('tar %s %s %s %s', $flags, $tarLocal, $excludes, $from);
 
         /** @var Process $process */
         $process = $this->runtime->runLocalCommand($cmdTar, 300);
@@ -51,7 +52,7 @@ class PrepareTask extends AbstractTask
 
     protected function getExcludes()
     {
-        $excludes = $this->runtime->getEnvOption('exclude', []);
+        $excludes = $this->runtime->getMergedOption('exclude', []);
         $excludes = array_merge(['.git'], array_filter($excludes));
 
         foreach ($excludes as &$exclude) {