+++ src/Task/BuiltIn/Symfony/AssetsInstallTask.php
@@ -11,14 +11,13 @@
 namespace Mage\Task\BuiltIn\Symfony;
 
 use Symfony\Component\Process\Process;
-use Mage\Task\AbstractTask;
 
 /**
  * Symfony Task - Install Assets
  *
  * @author Andrés Montañez <andresmontanez@gmail.com>
  */
-class AssetsInstallTask extends AbstractTask
+class AssetsInstallTask extends AbstractSymfonyTask
 {
     public function getName()
     {
@@ -41,14 +40,8 @@ class AssetsInstallTask extends AbstractTask
         return $process->isSuccessful();
     }
 
-    protected function getOptions()
+    protected function getSymfonyOptions()
     {
-        $options = array_merge(
-            ['console' => 'bin/console', 'env' => 'dev', 'target' => 'web', 'flags' => '--symlink --relative'],
-            $this->runtime->getMergedOption('symfony'),
-            $this->options
-        );
-
-        return $options;
+        return ['target' => 'web', 'flags' => '--symlink --relative'];
     }
 }