+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG for 3.X
 =================
 
+* 3.2.0 (2017-04-14)
+ * Allow to pre-register Custom Tasks
+ * [PR#365] New option "from" to define deployment start point
+ * Allow to define excludes in the global scope.
+ * Improve code quality, remove duplications on Symfony Tasks.
+ * Improve code quality, remove duplications on Composer Tasks.
+ * [PR#364] Allow to define custom timeout to Composer:Install
+
 * 3.1.0 (2017-02-25)
  * Add new Exec task to execute arbitrary shell commands
  * Add new Composer task, to update phar (composer/self-update)